requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==41.0.2
Django==4.1
django-redis==4.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
msgpack-python==0.5.6
psycopg2==2.9.6
pycparser==2.21
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3